+++ NEWS.md
@@ -1,3 +1,5 @@
+# agriutilities 1.1.0
+
 # agriutilities 1.0.0.9000
 
 ## Minor changes/fixes